refactored Board components into their own files
