+++ frontend/src/components/Board/BoardComponents/MemberBar.tsx
@@ -1,0 +1,5 @@
+
+export function MemberBar(props) {
+  return <h1>Hello, {props.name}</h1>;
+}
+
